feats: add more vim movements + complete core
Milestone commit: This commit completes the core
functionality of the editor. Specifically, the
`enter at middle of text` functionality has been implemented.
Also, this commit adds the `x` to delete, `Shift + i` to
insert at the beginning of the line, and `Shift + a` to
insert at the end of the line vim movements to Ronto.